requirements.txt
requests
beautifulsoup4
dnspython
builtwith
ipwhois
colorama
pyfiglet
